+++ plugins/scaffolder-actions/scaffolder-backend-module-utils/src/actions/jsonata/json.ts
@@ -93,7 +93,7 @@ export function createJsonJSONataTransformAction() {
 
       const data = JSON.parse(fs.readFileSync(sourceFilepath).toString());
       const expression = jsonata(ctx.input.expression);
-      const result = expression.evaluate(data);
+      const result = await expression.evaluate(data);
 
       ctx.output('result', resultHandler(result));
     },